util: moving http server code to util package.
The http server code used for CDC could be re-used for other sources. This change
moves the generic code into the util/stdserver package.
The generic http server will provide the common logic for authentication,
diagnostic, metrics, and secure transport.

Summary of changes:

- The reusable code in source/server/provider.go has moved to the following
  files under util/stdserver.
  auth.go: constructs an authenticator (JWT-based or no-op)
  listener.go: constructs the incoming network socket for the server.
  secure.go: provides tls configuration
  server.go: constructs an http server with the given capabilities (authentication,
             security etc.)

- The code specific go cdc in source/server has moved to source/cdc/server

- The code under staging/auth moved to util/auth to insure the util/stdserver only depends
  on util/types for internal packages.

Fixes: #597.
