+++ internal/util/stdserver/secure.go
@@ -1,0 +1,87 @@
+// Copyright 2023 The Cockroach Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package stdserver
+
+import (
+	"crypto/ecdsa"
+	"crypto/elliptic"
+	cryptoRand "crypto/rand"
+	"crypto/tls"
+	"crypto/x509"
+	"crypto/x509/pkix"
+	"math/big"
+	"net"
+	"time"
+
+	"github.com/pkg/errors"
+)
+
+// TLSConfig loads the certificate and key
+// from disk, to generate a self-signed localhost certificate, or to
+// return nil if TLS has been disabled.
+func TLSConfig(config *Config) (*tls.Config, error) {
+	if config.TLSCertFile != "" && config.TLSPrivateKey != "" {
+		cert, err := tls.LoadX509KeyPair(config.TLSCertFile, config.TLSPrivateKey)
+		if err != nil {
+			return nil, err
+		}
+		return &tls.Config{Certificates: []tls.Certificate{cert}}, nil
+	}
+
+	if !config.GenerateSelfSigned {
+		return nil, nil
+	}
+
+	// Loosely based on https://golang.org/src/crypto/tls/generate_cert.go
+	priv, err := ecdsa.GenerateKey(elliptic.P256(), cryptoRand.Reader)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to generate private key")
+	}
+
+	now := time.Now().UTC()
+
+	serialNumberLimit := new(big.Int).Lsh(big.NewInt(1), 128)
+	serialNumber, err := cryptoRand.Int(cryptoRand.Reader, serialNumberLimit)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to generate serial number")
+	}
+
+	cert := x509.Certificate{
+		BasicConstraintsValid: true,
+		DNSNames:              []string{"localhost"},
+		IPAddresses:           []net.IP{net.IPv4(127, 0, 0, 1), net.IPv6loopback},
+		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
+		KeyUsage:              x509.KeyUsageKeyEncipherment | x509.KeyUsageDigitalSignature,
+		NotBefore:             now,
+		NotAfter:              now.AddDate(1, 0, 0),
+		SerialNumber:          serialNumber,
+		Subject: pkix.Name{
+			Organization: []string{"Cockroach Labs"},
+		},
+	}
+
+	bytes, err := x509.CreateCertificate(cryptoRand.Reader, &cert, &cert, &priv.PublicKey, priv)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to generate certificate")
+	}
+
+	return &tls.Config{
+		Certificates: []tls.Certificate{{
+			Certificate: [][]byte{bytes},
+			PrivateKey:  priv,
+		}}}, nil
+}
