+++ internal/util/stdserver/wrapper.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package server
+package stdserver
 
 import (
 	"io"
